Search results is not shown in x86_64 machine.
Use ssize_t, size_t.

+++ src/filter.c
@@ -181,7 +181,7 @@ gchar *get_cache_file(gchar *path)
 	gint i;
 	gint r;
 	gchar *p;
-	gint length;
+	ssize_t length;
 	GError *error=NULL;
 	gchar *contents;
 	gchar tmpout[512];
