+++ src/jcode.c
@@ -49,11 +49,11 @@ static gchar *replace_table[] = {
 	NULL, NULL, NULL, NULL // 13-89 - 92
  };
 
-void replace_char(const guchar *icode, const guchar *ocode, guchar **inbuf, guchar **outbuf, gint *isize, gint *osize){
+void replace_char(const guchar *icode, const guchar *ocode, guchar **inbuf, guchar **outbuf, size_t *isize, size_t *osize){
 	guchar *in, *out;
 	guchar *str;
 	guchar *utf_str;
-	gint len;
+	size_t len;
 
 	in = *inbuf;
 	out = *outbuf;
